+++ bn_mp_read_radix.c
@@ -50,11 +50,11 @@ int mp_read_radix(mp_int *a, const char *str, int radix)
        * [e.g. in hex]
        */
       ch = (radix <= 36) ? (char)toupper((int)*str) : *str;
-      pos = ch - '(';
+      pos = (unsigned)(ch - '(');
       if (mp_s_rmap_reverse_sz < pos) {
          break;
       }
-      y = mp_s_rmap_reverse[pos];
+      y = (int)mp_s_rmap_reverse[pos];
 
       /* if the char was found in the map
        * and is less than the given radix add it
